error on display and filter cause lobstream

+++ src/main/java/com/Platinum/Asixstore/Controller/BarangController.java
@@ -3,6 +3,7 @@ package com.Platinum.Asixstore.Controller;
 import com.Platinum.Asixstore.Dto.BarangDto;
 import com.Platinum.Asixstore.Entity.Barang;
 import com.Platinum.Asixstore.Entity.User;
+import com.Platinum.Asixstore.Repository.BarangRepo;
 import com.Platinum.Asixstore.Repository.UserRepo;
 import com.Platinum.Asixstore.Service.BarangService;
 import net.bytebuddy.utility.nullability.AlwaysNull;
@@ -14,63 +15,76 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.transaction.Transactional;
 import java.io.IOException;
 import java.util.List;
+import java.util.Locale;
 
 @RestController
+@Transactional
 public class BarangController {
     @Autowired
     BarangService barangService;
     @Autowired
     UserRepo userRepo;
-    public Authentication authentication(){
+    @Autowired
+    BarangRepo barangRepo;
+
+    public Authentication authentication() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         return auth;
     }
+
     @PostMapping("/barang/{userId}/daftar")
     public ResponseEntity<?> submit_barang(@PathVariable int userId, BarangDto barangDto, @RequestParam("barangImg") MultipartFile fileUpload) throws IOException {
         User userToken = userRepo.findById(userId);
-        if (userToken.getEmail().equalsIgnoreCase(authentication().getPrincipal().toString())){
-            if (userId == barangDto.getUserId()){
-                barangDto.setBarangImg(fileUpload);
-                Barang barang = barangService.submit_barang(userId,barangDto);
-                return new ResponseEntity<>(barang, HttpStatus.CREATED);
-            }else {
-                String error = "login error cok";
-                return new ResponseEntity<>(error,HttpStatus.ACCEPTED);
-            }
-        }else {
+        if (userToken.getEmail().equalsIgnoreCase(authentication().getPrincipal().toString())) {
+            barangDto.setBarangImg(fileUpload);
+            Barang barang = barangService.submit_barang(userId, barangDto);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        } else {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
     }
+
     @GetMapping("/barang")
     public ResponseEntity<?> display_barang() {
-        List<Barang> barangs = barangService.display_barang();
-        return new ResponseEntity<>(barangs, HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(barangService.display_barang(), HttpStatus.ACCEPTED);
     }
 
     @RequestMapping(value = "/barang/{tipeBarang}", method = RequestMethod.GET)
     public ResponseEntity<?> filter_barang(@PathVariable("tipeBarang") String tipeBarang) throws Exception {
-        List<Barang> barangFilter = barangService.filter_barang(tipeBarang);
+        List<Barang> barangFilter = barangService.filter_barang(tipeBarang.toLowerCase(Locale.ROOT));
         return new ResponseEntity<>(barangFilter, HttpStatus.ACCEPTED);
     }
+
     @PutMapping("/barang/update/{barangId}")
-    public ResponseEntity<?> beli_tawar_harga(@PathVariable("barangId") int barangId,BarangDto barangDto) {
-        Barang barang = barangService.update_harga_tawar(barangId,barangDto);
+    public ResponseEntity<?> beli_tawar_harga(@PathVariable("barangId") int barangId, BarangDto barangDto) {
+        Barang barang = barangService.update_harga_tawar(barangId, barangDto);
         return new ResponseEntity<>(barang, HttpStatus.ACCEPTED);
     }
-    @DeleteMapping("/barang/delete/{barangId}")
-    public ResponseEntity<?> hapus_barang(@PathVariable("barangId") int barangId){
-        boolean barang_status = barangService.delete_barang(barangId);
-        if (barang_status){
-            return new ResponseEntity<>(barang_status, HttpStatus.ACCEPTED);
-        }else {
-            return new ResponseEntity<>(barang_status, HttpStatus.NO_CONTENT);
+
+    @DeleteMapping("barang/delete/{barangId}")
+    public ResponseEntity<?> hapus_barang(@PathVariable("barangId") int barangId) {
+        Barang barang = barangRepo.findByBarangId(barangId);
+        User user = userRepo.findById(barang.getUser().getUserId());
+
+
+        if (authentication().getPrincipal().toString().equalsIgnoreCase(barang.getUser().getEmail())) {
+            boolean barang_status = barangService.delete_barang(barangId);
+            if (barang_status) {
+                return new ResponseEntity<>(barang_status, HttpStatus.ACCEPTED);
+            } else {
+                return new ResponseEntity<>(barang_status, HttpStatus.NO_CONTENT);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
+
     @PutMapping("/barang/update/{userId}/{barangId}")
-    public ResponseEntity<?> update_barang(@PathVariable("barangId") int barangId,@PathVariable("userId") int userId, @RequestParam ("barangImg") MultipartFile fileUpload, BarangDto barangDto)throws  IOException{
+    public ResponseEntity<?> update_barang(@PathVariable("barangId") int barangId, @PathVariable("userId") int userId, @RequestParam("barangImg") MultipartFile fileUpload, BarangDto barangDto) throws IOException {
         barangDto.setBarangImg(fileUpload);
         barangService.edit_barang(barangId, userId, barangDto);
         Barang response = barangService.display_barang_byId(barangId);
